api/pyproject.toml
[project]
name = "zero-click-api"
version = "1.0.0"
requires-python = ">=3.10,<3.14"
dependencies = [
    "fastapi>=0.128.0",
    "uvicorn[standard]>=0.40.0",
    "requests>=2.32.0",
    "google-cloud-firestore>=2.22.0",
    "google-api-python-client>=2.187.0",
    "google-auth>=2.37.0",
    "pydantic>=2.12.5",
    "carconnectivity>=0.10",
    "carconnectivity-connector-audi>=0.2.0",
    "carconnectivity-connector-volkswagen>=0.1.0",
    "carconnectivity-connector-skoda>=0.1.0",
    "carconnectivity-connector-seatcupra>=0.1.0",
    "renault-api>=0.2.0",
    "aiohttp>=3.9.0",
    "TeslaPy>=2.8.0",
    "PyJWT[crypto]>=2.10.1",
    "google-cloud-secret-manager>=2.26.0",
    "google-cloud-kms>=3.7.0",
    "slowapi>=0.1.9",
    "cryptography>=44.0.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
python_files = ["test_*.py"]
python_functions = ["test_*"]
